refactor(executable): move lib.ts to utils/index.ts
The lib.ts file has been moved to a new directory called utils. This
change is part of a larger refactoring effort to improve the structure
and organization of the codebase.
